Initial SQLite database build integration
Added the SQLite database for the geo-features to the build and added just the code to open a connection to the database and then close it.  I'll fill in the code to actually query the database and create the map Polylines and POIs later; I just wanted to get the build process under control.