Remove the GoogleMaps.framework from the repository
Remove the GoogleMaps.framework from the repository; it was stupid to have
checked it in.  Now there is a custom build stage to check for the presence of
the framework, and unpack it into place if it isn't there.  The script will
search through the project directory, the parent directory above the project,
the user's Desktop, and then the user's Home, in that order, for the .zip file
as Google distributes it.  If it is not found any of those places, it will be
downloaded automatically, and left in the parent directory of the repository.
Hopefully this is a good compromise on only slowing down the build the first
time after a repositiory is checked out, and even if different copies of the
repository are checked out (either because the original one was deleted, or
because a second work area is desired), the download will only have to happen
once.