Remove the Google Maps API key from the source
The key is now configured during the first build, at the same time as the
GoogleMaps.framework is unpacked into the project directory.  If the file
GoogleMaps-API-Key.txt exists along side the .zip file for the Google Maps SDK,
then the key will be inserted into the code before being built.  If the file
does not exist, the user will be prompted to select the key and copy it to the
clipboard and then hit the Okay button.  At that point, the key file will be
created (so that the process will only ever need to be done once), and the key
will be inserted into the code before being built.

The previous key is going to get invalidated, and when I generate a new one,
I'll send it out to the team by email.