Updates for the auto-tracking
Added some updates to get the auto-tracking done right, but it's currently
disabled (commented out the actual zooming) since I don't want it dragging away
from the actual map, since none of us are actually there.  Also turned on the
"myLocation" and compass buttons on the map.  Not sure if the myLocation button
is actually working, since it goes to where the code was auto-zooming anyway.
I need to get a check-box preference in place so the auto-follow can be turned
on and off dynamically.