Got the custom pop-ups working
Figured out the reason that my custom handler for generating the pop-ups for
POIs wasn't working, and roughed in the generation of the pop-up.  The bubble
auto-sizes itself for the size of the title, but not smaller than the size of
the buttons below it.  The buttons are the "More info" button and the "Share"
button.  The "Share" button is set up to be automatically disable-able (the
config code isn't there to actually make it dynamic yet), and not having the
button will resize everything accordingly.  The target of those two buttons is
not in place yet, so the "More info" button target is the next thing to put in.