Added the ConfigModel class for configuration management and save/restore
Added the ConfigModel class to the build and wove it in to the fabric of the
AppDelegate, which passes it to the MapViewController.  When I get the
ConfigViewController put together, it'll get passed in to there to be updated,
and hopefully when the MapViewController is loaded again, it will do its bit of
gathering config from the ConfigModel again.

The ConfigModel has code to serialize its data, which it writes to the
UserDefaults config area, so it's stored on disk as a .plist.