Minor enhancement to Install_GMapsAPI.sh script
I just added the --key and --unkey command line options to the script so that
it is more convenient to add or remove the Google Maps API key from
AppDelegate.m, which one might want to do before a pull/merge that includes
changes to AppDelegate.m.  It is still called by the build stages
automatically, and will only add the key when called without options if it is
also installing the GoogleMaps.framework; this might be desirable to change,
but for now I'm keeping it completely identical in behavior to what it was
before.  The --key option, though will force the updating, but only from the
unkeyed state (as AppDelegate.m is in the state committed to the repository),
and it will do it from the cached version (which is stored along side where
ever it finds the .zip file for the framework.)  The --unkey flag should leave
it in the state that it should be in before commiting any changes to
AppDelegate.m to the repository.