Switched the ConfigModel to use the singleton pattern for convenience
Made the ConfigModel object a singleton, rather than passing it around between
the UIViewControlers.  That looks like it would have been do-able, but I
probably should have just made it a singleton to begin with, so that I don't
have to wire it expressly to the UINavController, and then send it in to each
UIViewController when they are put on top of the view stack.