Added the SettingsViewController and the map type option
I tried to minimize the amount of changes to the story board to just what I
needed to get the code working; hopefully it doesn't screw anything else up.

I added the SettingsViewController class and set it up so that the map type
could be selected for the main map.