Fixed up the ConfigModel initialization from the stored state.
Removed the instantiation of the ConfigModel singleton from AppDelegate, which
was also where it was being loaded from the UserDefaults, because I suspect
that AppDelegate.m is running in a different thread than the UIViewControllers,
which is where I actually need them.  Now the ConfigModel singleton
automatically loads itself from UserDefaults if it is creating the singleton in
getConfigModel.  Now everywhere that needs it will get the one available in the
current thread, initialized properly, and any time that changes are made, the
changes should be saved right then.