Factored out the map object creation
Since I want to re-draw the map objects when we come back to the
MapViewController, I factored out the code that queries the database and
creates the map objects for overlay on the map.  Now that code is called again
every time that the viewWillAppear method is called by the system.  This means
that if changes are made to the filter, the redrawing will happen
automatically.