Major fixes/re-write of the POI marker info view bubbles
Well, it turns out that the custom POI info view bubble that I coded looked
good, and had the buttons that Google won't allow you to have normally, but
since I hadn't hooked anything up to the button yet, I didn't realize that
apparently the hook that Google gives you to make your own custom info bubbles
takes your custom contents and basically takes a picture of it, and puts that
in the bubble.  The end result is that the buttons weren't really buttons that
could be tapped.  So, I had to make an amazingly ugly hack, where I produce my
custom content, and then create a phantom UIView that is the same size, but
empty, and embed that in my own simulation of the bubble that Google draws
(they provide that hook, should you be so desperate...)  Then, I have to figure
out where on the screen that is, and place my custom content UIView directly on
top of it, which is a bit of a pain because when the user drags the map, you
have to make sure the UIView that you put on top moves around to stay on top.
Also, when that bubble disappears, you have to make sure you remove the
overlayed UIView as well, and there are quite a few places that you have to
hook that.

End result, the WebView that I added to the "POI Detail View" scene, which is
included in this commit, is finally ready for me to try to push a URL to it and
programmatically segue to it, if I can figure out how to do that.  But the
button that will initiate that seems to finally be working, after a few hours
of wasted time screwing around trying to get something to work that I thought
was already done!