Got the POI Detail View scene working
It's a bit complicated about how the transition from the Map scene to the POI
Detail View scene works, because it's programmatic, and it needs data to go
with it (the URL), but I've got it ironed out.  This commit has the newly
created WebKitViewController class, which just waits until it is started, and
then loads what ever URL is in the url property.  The MapViewController goes
fishing when it first loads to find the WebKitViewController through the
UIStoryboard object (the storyboard item has an identifier string of @"POI
Detail View" to facilitate this lookup.)  Then when the button is pushed in the
custom marker info pop-up, a callback is made to a glue function that just sets
the url property in the WebKitViewController and then pushes it on the
UINavigationController's view controller stack.  This gives us the automatic
"Back" button to take us back to the map.